+++ src/dev/flang/ast/ParsedCall.java
@@ -26,7 +26,11 @@ Fuzion language implementation.  If not, see <https://www.gnu.org/licenses/>.
 
 package dev.flang.ast;
 
+import java.util.ListIterator;
+
+import dev.flang.util.FuzionConstants;
 import dev.flang.util.List;
+import dev.flang.util.SourcePosition;
 
 
 /**
@@ -34,6 +38,10 @@ import dev.flang.util.List;
  */
 public class ParsedCall extends Call
 {
+
+
+  final List<Expr> _parsedActuals;
+  boolean _appliedPartially = false;
 
   /**
    * Constructor to read a field in target t
@@ -45,6 +53,8 @@ public class ParsedCall extends Call
   public ParsedCall(Expr target, ParsedName name)
   {
     super(name._pos, target, name._name);
+
+    _parsedActuals = NO_PARENTHESES;// Expr.NO_EXPRS;
   }
 
 
@@ -60,7 +70,25 @@ public class ParsedCall extends Call
    */
   public ParsedCall(Expr target, ParsedName name, List<Expr> arguments)
   {
-    super(name._pos, target, name._name, arguments);
+    this(name._pos, target, name, arguments);
+  }
+
+
+  /**
+   * Constructor to call feature with name 'n' on target 't' with actual
+   * arguments 'la'.
+   *
+   * @param target the target of the call, null if none.
+   *
+   * @param name the name of the called feature
+   *
+   * @param arguments list of actual arguments
+   */
+  public ParsedCall(SourcePosition pos, Expr target, ParsedName name, List<Expr> arguments)
+  {
+    super(pos, target, name._name, arguments);
+
+    _parsedActuals = arguments;
   }
 
 
@@ -78,16 +106,32 @@ public class ParsedCall extends Call
   public ParsedCall(Expr target, ParsedName name, int select)
   {
     super(name._pos, target, name._name, select, NO_PARENTHESES);
+
+    _parsedActuals = NO_PARENTHESES;
   }
 
 
   boolean isInfixPipe(boolean parenthesesAllowed)
   {
-    return isOperatorCall(parenthesesAllowed) && name().equals("infix |") && _actualsNew.size() == 1;
+    return isOperatorCall(parenthesesAllowed) && name().equals("infix |") && _parsedActuals.size() == 1;
   }
 
 
+  /**
+   * True iff this call was performed giving 0 or more actual arguments in
+   * parentheses.  This allows a distinction between "a.b" and "a.b()" if b has
+   * no formal arguments and is of a fun type. In this case, "a.b" calls only b,
+   * while "a.b()" is syntactic sugar for "a.b.call".
+   *
+   * @return true if parentheses were present.
+   */
+  public boolean hasParentheses()
+  {
+    return _appliedPartially || _parsedActuals != NO_PARENTHESES;
+  }
 
+
+  @Override
   public ParsedType asParsedType()
   {
     var target = target();
@@ -111,7 +155,7 @@ public class ParsedCall extends Call
             name = "choice";
             tt = null;
           }
-        for (var a : _actualsNew)
+        for (var a : _parsedActuals)
           {
             var at = a.asParsedType();
             ok = ok && at != null;
@@ -122,6 +166,188 @@ public class ParsedCall extends Call
               : null;
   }
 
+
+  /*-------------------------------------------------------------------*/
+
+
+  /**
+   * Check if this call is a chained boolean call of the form
+   *
+   *   b <= c < d
+   *
+   * or, if the LHS is also a chained bool
+   *
+   *   (a < {t1 := b; t1} && t1 <= c) < d
+   *
+   * and return the part of the LHS that has the term that will need to be
+   * stored in a temp variable, 'c', as an argument, i.e., 'b <= c' or 't1 <=
+   * c', resp.
+   *
+   * @return the term whose RHS would have to be stored in a temp variable for a
+   * chained boolean call.
+   */
+  private Call chainedBoolTarget(Resolution res, AbstractFeature thiz)
+  {
+    Call result = null;
+    if (Types.resolved != null &&
+        targetFeature(res, thiz) == Types.resolved.f_bool &&
+        isInfixOperator() &&
+        target() instanceof ParsedCall pc &&
+        pc.isInfixOperator() &&
+        pc.isOperatorCall(false))
+      {
+        result = (pc._actuals.get(0) instanceof Call acc && acc.isChainedBoolRHS())
+          ? acc
+          : pc;
+      }
+    return result;
+  }
+
+
+  /**
+   * if loadCalledFeature is about to fail, try if we can convert this call into
+   * a chain of boolean calls:
+   *
+   * check if we have a call of the form
+   *
+   *   a < b <= c
+   *
+   * and convert it to
+   *
+   *   a < {tmp := b; tmp} && tmp <= c
+   */
+  @Override
+  protected void findChainedBooleans(Resolution res, AbstractFeature thiz)
+  {
+    var cb = chainedBoolTarget(res, thiz);
+    if (cb != null && _actuals.size() == 1)
+      {
+        var b = res.resolveType(cb._actuals.getLast(), thiz);
+        if (b.typeForInferencing() != Types.t_ERROR)
+          {
+            String tmpName = FuzionConstants.CHAINED_BOOL_TMP_PREFIX + (_chainedBoolTempId_++);
+            var tmp = new Feature(res,
+                                  pos(),
+                                  Visi.PRIV,
+                                  b.type(),
+                                  tmpName,
+                                  thiz);
+            Expr t1 = new Call(pos(), new Current(pos(), thiz), tmp, -1);
+            Expr t2 = new Call(pos(), new Current(pos(), thiz), tmp, -1);
+            var movedTo = new Call(pos(), t2, name(), _parsedActuals)
+              {
+                boolean isChainedBoolRHS() { return true; }
+              };
+            this._movedTo = movedTo;
+            Expr as = new Assign(res, pos(), tmp, b, thiz);
+            t1 = res.resolveType(t1    , thiz);
+            as = res.resolveType(as    , thiz);
+            var result = res.resolveType(movedTo, thiz);
+            cb._actuals.set(cb._actuals.size()-1,
+                            new Block(new List<Expr>(as, t1)));
+            _actuals = new List<Expr>(result);
+            _calledFeature = Types.resolved.f_bool_AND;
+            _pendingError = null;
+            _name = _calledFeature.featureName().baseName();
+          }
+      }
+  }
+
+
+  /**
+   * Does this call a non-generic infix operator?
+   */
+  private boolean isInfixOperator()
+  {
+    return
+      _name.startsWith("infix ") &&
+      (_actuals.size() == 1 /* normal infix operator 'a.infix + b' */                ||
+       _actuals.size() == 2 /* infix on different target 'X.Y.Z.this.infix + a b' */    ) &&
+      true; /* no check for _generics.size(), we allow infix operator to infer arbitrary number of type parameters */
+  }
+
+
+  @Override
+  protected void splitOffTypeArgs(Resolution res, AbstractFeature calledFeature, AbstractFeature outer)
+  {
+    var g = new List<AbstractType>();
+    var a = new List<Expr>();
+    var ts = calledFeature.typeArguments();
+    var tn = ts.size();
+    var ti = 0;
+    var vs = calledFeature.valueArguments();
+    var vn = vs.size();
+    var i = 0;
+    ListIterator<Expr> ai = _actuals.listIterator();
+    while (ai.hasNext())
+      {
+        var aa = ai.next();
+
+        // check that ts[ti] is open type parameter only iff ti == tn-1, ie.,
+        // only the last type parameter may be open
+        if (CHECKS) check
+          (ti >= tn-1 ||
+           ts.get(ti).kind() == AbstractFeature.Kind.TypeParameter    ,
+           ti != tn-1 ||
+           ts.get(ti).kind() == AbstractFeature.Kind.TypeParameter     ||
+           ts.get(ti).kind() == AbstractFeature.Kind.OpenTypeParameter);
+
+        if (_actuals.size() - i > vn)
+          {
+            AbstractType t = _parsedActuals.get(i).asParsedType();
+            if (t != null)
+              {
+                g.add(t);
+              }
+            ai.set(Expr.NO_VALUE);  // make sure visit() no longer visits this
+            if (ts.get(ti).kind() != AbstractFeature.Kind.OpenTypeParameter)
+              {
+                ti++;
+              }
+          }
+        else
+          {
+            a.add(aa);
+          }
+        i++;
+      }
+    _generics = g;
+    _actuals = a;
+  }
+
+
+  @Override
+  protected Call resolveImmediateFunctionCall(Resolution res, AbstractFeature outer)
+  {
+    Call result = this;
+
+    // replace Function or Lazy value `l` by `l.call`:
+    if (isImmediateFunctionCall())
+      {
+        result = pushCall(res, outer, "call").resolveTypes(res, outer);
+      }
+    return result;
+  }
+
+
+  /**
+   * Is this call returning a Function/lambda that should
+   * immediately be called?
+   */
+  private boolean isImmediateFunctionCall()
+  {
+    return
+      _type.isFunctionType()                      &&
+      _calledFeature != Types.resolved.f_Function && // exclude inherits call in function type
+      _calledFeature.arguments().size() == 0      &&
+      hasParentheses()
+      ||
+      _type.isLazyType()                          &&   // we are `Lazy T`
+      _calledFeature != Types.resolved.f_Lazy     &&   // but not an explicit call to `Lazy` (e.g., in inherits clause)
+      _calledFeature.arguments().size() == 0      &&   // no arguments (NYI: maybe allow args for `Lazy (Function R V)`, then `l a` could become `c.call.call a`
+      _parsedActuals.isEmpty()                    &&   // dto.
+      originalLazyValue() == this;                     // prevent repeated `l.call.call` when resolving the newly created Call to `call`.
+  }
 
 }
